+++ src/runtime-rs/crates/hypervisor/src/dragonball/mod.rs
@@ -16,12 +16,17 @@ use std::sync::Arc;
 
 use anyhow::{Context, Result};
 use async_trait::async_trait;
+use dbs_utils::net::MacAddr as DragonballMacAddr;
+use dragonball::api::v1::{
+    Backend as DragonballNetworkBackend, NetworkInterfaceConfig as DragonballNetworkConfig,
+    VirtioConfig as DragonballVirtioConfig,
+};
 use kata_types::capabilities::Capabilities;
 use kata_types::config::hypervisor::Hypervisor as HypervisorConfig;
 use tokio::sync::RwLock;
 use tracing::instrument;
 
-use crate::{DeviceType, Hypervisor, VcpuThreadIds};
+use crate::{DeviceType, Hypervisor, NetworkBackend, NetworkConfig, VcpuThreadIds};
 
 pub struct Dragonball {
     inner: Arc<RwLock<DragonballInner>>,
@@ -188,5 +193,49 @@ impl Persist for Dragonball {
         Ok(Self {
             inner: Arc::new(RwLock::new(inner)),
         })
+    }
+}
+
+impl From<NetworkBackend> for DragonballNetworkBackend {
+    fn from(value: NetworkBackend) -> Self {
+        match value {
+            NetworkBackend::Virtio(config) => Self::Virtio(DragonballVirtioConfig {
+                iface_id: config.virt_iface_name.clone(),
+                host_dev_name: config.host_dev_name.clone(),
+                rx_rate_limiter: None,
+                tx_rate_limiter: None,
+                allow_duplicate_mac: config.allow_duplicate_mac,
+            }),
+            NetworkBackend::Vhost(config) => Self::Vhost(DragonballVirtioConfig {
+                iface_id: config.virt_iface_name.clone(),
+                host_dev_name: config.host_dev_name.clone(),
+                rx_rate_limiter: None,
+                tx_rate_limiter: None,
+                allow_duplicate_mac: config.allow_duplicate_mac,
+            }),
+        }
+    }
+}
+
+impl From<NetworkConfig> for DragonballNetworkConfig {
+    fn from(value: NetworkConfig) -> Self {
+        let r = &value;
+        r.into()
+    }
+}
+
+impl From<&NetworkConfig> for DragonballNetworkConfig {
+    fn from(value: &NetworkConfig) -> Self {
+        Self {
+            num_queues: Some(value.queue_num),
+            queue_size: Some(value.queue_size as u16),
+            backend: value.backend.clone().into(),
+            guest_mac: value.guest_mac.clone().map(|mac| {
+                // We are safety since mac address is checked by endpoints.
+                DragonballMacAddr::from_bytes(&mac.0).unwrap()
+            }),
+            use_shared_irq: value.use_shared_irq,
+            use_generic_irq: value.use_generic_irq,
+        }
     }
 }
